Training startup updates   - Remove alignments path option from cli   - Restructure training_data.py   - Load alignment data from PNG header during first epoch   - lib.image.read_image_batch - Add option to return metadata   - lib.utils.get_image_paths - Add option for explicit extension   - plugins.train.trainer._base - remove pre-cache alignments code   - scripts.train - Check first image in training folders for metadata   - Documentation
