+++ docs/full/lib/training.rst
@@ -1,0 +1,24 @@
+****************
+training package
+****************
+
+The training Package handles the processing of faces for feeding into a Faceswap model.
+
+.. contents:: Contents
+   :local:
+
+augmentation module
+===================
+
+.. automodule:: lib.training.augmentation
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+generator module
+================
+
+.. automodule:: lib.training.generator
+   :members:
+   :undoc-members:
+   :show-inheritance:
