--- docs/full/lib/training_data.rst
@@ -1,18 +1,0 @@
-*********************
-training\_data module
-*********************
-
-.. rubric:: Module Summary
-
-.. autosummary::
-   :nosignatures:
-   
-   ~lib.training_data.ImageAugmentation
-   ~lib.training_data.TrainingDataGenerator
-
-.. rubric:: Module
-
-.. automodule:: lib.training_data
-   :members:
-   :undoc-members:
-   :show-inheritance:
