+++ lib/training/__init__.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python3
+""" Package for handling alignments files, detected faces and aligned faces along with their
+associated objects. """
+
+from .augmentation import ImageAugmentation
+from .generator import TrainingDataGenerator
